Moved /Hobo and /aquaMeaure inside functions so the same path can be set for compile_HOBO_data() and compile_aquaMeasure_data()
Added if()else() to convert_tidydata() for parsing DATES to be compatable with manual extractions and string package extractions

If the date is stored in excel as a number (manual extraction) janitor::convert_to_date_time is used. If not (stored as a character, eg from  string package), lubridate::parse_date_time is used.

Used same if()else() statement in compile_aquaMeasure_data() for consistency