Can now read in .csv or .xlsx (must be specified in args) We are going to start exporting HOBO data in csv format because it is in true UTC.
Removed tz.UTC from argumnets because it only applied to the xlsx files. Now if file.type == "xlsx", tz.UTC is automatically set to TRUE, and is FALSE otherwise. As we convert our workflow to csv files, the convert_to_real_UTC() function will be removed

Some files have Plot Title as the first line. If so, function will remove (to be more flexible in case some files have it and others dont)

Added in a format to parse_date_time so it can read in dates exported directly from HOBOware default. May need to add others